requirements.txt
sqlalchemy-iris
flask
flask_sqlalchemy
requests
openai